service/lesson_format_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
python-docx==1.1.0
markdown==3.5.2
html2text==2020.1.16
weasyprint==60.2
requests==2.31.0
pydantic==2.6.0
python-dotenv==1.0.0
pdfplumber==0.10.3
beautifulsoup4==4.12.3
